test(davinci): keep p2-12b budget tests under length limit
