+++ crates/vize/src/commands/davinci_ice.rs
@@ -45,6 +45,10 @@ use vize_davinci::pass::{
 };
 use vize_s0::{FxHashMap, String, cstr};
 
+mod budget;
+
+pub(crate) use budget::plan_budget;
+
 /// `artifact-stage` value for an embedded authored source.
 pub(crate) const ARTIFACT_STAGE_SOURCE: &str = "source";
 /// `[repro.config]` key naming the backend mode (`dom`/`ssr`/`vapor`).
@@ -82,19 +86,6 @@ pub(crate) fn compile_plan(ssr: bool, vapor: bool) -> (&'static Pipeline, &'stat
     } else {
         (&legacy_plan::DOM, "dom")
     }
-}
-
-/// Measure the selected build plan through the P2-3 budget observer.
-///
-/// This is intentionally plan-level accounting: today the real build stages
-/// are still the legacy compile path, so the observer reports planned fusion
-/// groups for the selected backend. When the build path is routed through the
-/// pass manager, this helper remains the place that turns observer output into
-/// build-profile counters.
-pub(crate) fn plan_budget(plan: &Pipeline) -> BudgetObserver {
-    let mut budget = BudgetObserver::new();
-    run_pipeline(plan, &mut budget, |_| Ok(())).expect("compile plans use no-op pass bodies");
-    budget
 }
 
 /// `(ssr, vapor)` for a recorded mode name; `None` for an unknown mode.
@@ -260,55 +251,6 @@ pub(crate) fn source_repro(
     };
     folio.normalize();
     folio
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{compile_plan, plan_budget};
-    use vize_davinci::pass::{Fusability, PassDesc, PassKind, Pipeline, Preserved};
-
-    #[test]
-    fn compile_plan_budget_matches_selected_backend_groups() {
-        for (ssr, vapor, expected_mode) in [
-            (false, false, "dom"),
-            (true, false, "ssr"),
-            (false, true, "vapor"),
-        ] {
-            let (plan, mode) = compile_plan(ssr, vapor);
-            let budget = plan_budget(plan);
-
-            assert_eq!(mode, expected_mode);
-            assert_eq!(budget.pipelines, 1);
-            assert_eq!(budget.walks as usize, plan.group_count());
-            assert_eq!(budget.passes as usize, plan.passes.len());
-            assert_eq!(budget.failures, 0);
-        }
-    }
-
-    #[test]
-    fn plan_budget_counts_a_fused_build_group_once() {
-        const PASSES: &[PassDesc] = &[
-            PassDesc::new(
-                "facts",
-                PassKind::Optional,
-                Fusability::Fusable,
-                Preserved::ALL,
-            ),
-            PassDesc::new(
-                "emit",
-                PassKind::Optional,
-                Fusability::Fusable,
-                Preserved::ALL,
-            ),
-        ];
-        const PLAN: Pipeline = Pipeline::new("s2", PASSES);
-
-        let budget = plan_budget(&PLAN);
-
-        assert_eq!(budget.walks, 1);
-        assert_eq!(budget.passes, 2);
-        assert_eq!(budget.fusion_ratio_hundredths(), Some(200));
-    }
 }
 
 /// Write `folio` as `{stem}.repro.folio` under `dir`, creating `dir` first
